removing conftest
